bot-assistant/bot-assistant/requirements.txt
python-telegram-bot==21.5
httpx==0.27.2
google-api-python-client==2.151.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
Pillow==11.0.0
imageio-ffmpeg==0.6.0

